server: Normalize plugin can_* values
The documentation for .can_write and friends mentioned merely that the
callback must return a boolean value, or -1 on failure.  Historically,
we have treated any non-zero value other than -1 as true, and some
plugins have sloppily relied on this behavior, including the
standalone nbd plugin prior to the use of libnbd in v1.14.  But in
1.15.1, we added an assert in our backend code if an enabled feature
bit is not exactly 1.  To avoid tripping the assertion, we have to
normalize the plugin values before storing our internal value.

Fixes: c306fa93ab
Reported-by: Svenne Krap (https://svenne.dk)
Signed-off-by: Eric Blake <eblake@redhat.com>