tests: Add test to cover unusual .can_flush return
We want some testsuite coverage of handling non-1 .can_flush values.
The only in-tree plugin with this property historically was
nbd-standalone, but it didn't get frequently tested, and was recently
changed to no longer have that property (commit 15d4c4f1).  Revive the
specific behaviors we want to guarantee into a new independent test.

Signed-off-by: Eric Blake <eblake@redhat.com>