sh, eval: Cache .can_zero and .can_flush
In commit c306fa93ab and neighbors (v1.15.1), a concerted effort went
into caching the results of .can_FOO callbacks, with commit messages
demonstrating that a plugin with a slow callback should not have that
delay magnified multiple times.  But nothing was added to the
testsuite at the time, and with the sh and eval plugins, we still have
two scenarios where we cannot take advantage of the nbdkit cache
because we directly call things ourselves (.can_fua pre-existing since
v1.13.9, .can_fast_zero introduced in v1.15.1 right after the
cleanups) - because our language bindings unconditionally provide
.can_FOO, they have to duplicate logic that nbdkit would do for a C
plugin where .can_FOO is absent.  Fix the double call to .can_zero and
.can_flush by expanding the struct containing the handle we pass to
nbdkit, in order to cache what we need. Then enhance test-eflags.sh to
ensure we don't regress again.

Note that nbdkit does not yet cache .thread_model; that will be
addressed in the next patch.  Furthermore, we end up duplicating the
caching that nbdkit itself does (just like we are already duplicating
logic), because it would be a layering violation for us to have enough
information to call into backend_can_flush().

Fixes: 05c5eed6f2
Fixes: 8490694d08
Signed-off-by: Eric Blake <eblake@redhat.com>