server: Better caching of .thread_model
Repeatedly calling out to the plugin to determine the thread model is
not only potentially expensive (for a plugin that takes its time
answering, as is easy to do with 'nbdkit eval thread_model="sleep 1;
echo parallel" ...'), but risks confusion if the plugin changes its
answer over time.  Expose the cache variable in locks.c to the rest of
the server code so that we can learn the thread model exactly once,
with testsuite coverage to ensure we don't regress.

Signed-off-by: Eric Blake <eblake@redhat.com>