python: Smarter default for .can_fast_zero
Similarly to the sh plugin, we want to let nbdkit advertise fast zero
support when it will not be calling into our .zero.  And just like the
sh plugin, this requires that we cache a bit of information ourselves
(since it would be a layering violation to get what nbdkit has itself
cached).  Thankfully, our defaults for .can_fua are easier implement
correctly than what the sh plugin has to do.

To demonstrate this, try adding:

def can_zero(h):
    print("can_zero")
    return False

to example.py, before running

./nbdkit python plugins/python/example.py --run 'qemu-nbd --list'

Signed-off-by: Eric Blake <eblake@redhat.com>