nbd: Drop nbd-standalone fallback
When we first used libnbd (nbdkit 1.14), libnbd v1.0 was brand new
(both were released on the same day), so we left in the ability to
build a limited nbd plugin without the use of libnbd for comparison,
and to give distros time to pick up libnbd.  But now that libnbd has
had more time under its belt, and we have learned that our
nbd-standalone code is not getting much testing, it's time to follow
through with our plan to finish off the old code, and now build the
nbd plugin only when libnbd is present.

Note that ./configure --disable-nbd-plugin is now synonymous with
./configure --without-libnbd; but I left both spellings working for
the time being.  In the future, we may want to remove
--disable-nbd-plugin (reverting commit 0a190bd4).

Signed-off-by: Eric Blake <eblake@redhat.com>