tests: Keep $port global
When adding 'local' to various shell variables, I failed to read the
comments that $port must be global (otherwise test-long-name.sh and
others fail).  But I didn't notice the problem until later, because we
were missing a Makefile dependency that regenerates functions.sh from
functions.sh.in before running any .sh test, which meant my initial
testing of that patch was done with a stale functions.sh that did not
have the problem.

Fixes: dd3e149913
Signed-off-by: Eric Blake <eblake@redhat.com>