tests: Swap nbdkit process order in test-nbd-tls-psk.sh
We're still seeing sporadic failures of 'nbdkit nbd tls=' (or hangs,
prior to commit d6ee94d1), and I'm still trying to come up with a root
cause fix (it may involve smarter use of gnutls_bye() in libnbd).  In
the meantime, here's what we know: sending SIGTERM to the 'nbdkit
example1' server can sometimes hit a race where the 'nbdkit nbd tls='
client process is stuck in a poll() on non-blocking fds waiting to see
EOF from the server, while the 'nbdkit example1' server process is
stuck in a recv() on blocking fds waiting to see if the client will do
a clean shutdown of the gnutls session, deadlocking both processes in
.close without either actually closing the underlying socket.

Rearranging the order in which we call our start_nbdkit function
changes things so that we instead send SIGTERM to the client process
first, and since we are more confident in nbdkit as server reacting to
a disappearing client than we are to libnbd reacting as client to a
disappearing server, we are less likely to hit the hang.  Reordering
the startups is also a chance for us to add our first testsuite
coverage of the 'nbd retry=' parameter added back in commit 0bb76bc7.

Signed-off-by: Eric Blake <eblake@redhat.com>