nbd: Don't reference stale errno in reader loop
When switching to libnbd in commit ab7760fcfd, I assumed that after a
POLLIN event fires on the pipe-to-self, then read() will return either
1 or -1.  In practice, this assumption worked because POLLIN should
not fire unless there is data, and a single read will either find that
data or hit an (unlikely) error.  But in general, it is a latent bug,
because read() can also return 0 if the pipe hits EOF, in which case
POSIX says errno has an unspecified value, and we should not be
deciding whether to log an error based on a random value; the only
reason we didn't hit EOF is because we didn't loop the read() nor
close() the write end of the pipe until after the reader loop is
joined.

Later, when refactoring the pipe-to-self to be non-blocking, I still
did not adjust the assumption, but made the code harder to read with a
loop body of ';' buried after a comment.  That approach also made the
code less efficient in that it requires at least two read() calls per
POLLIN even when there is no backlog, and where the second read() will
(usually) fail with EAGAIN because we drained available data.  At any
rate, it was still relying on the assumption that a return of 0 will
not happen, which is a latent risk.

It is already hard to create a backlog of data in the pipe-to-self,
but even when one exists, it's worth recognizing that our inner loop
to drain the pipe-to-self will still be reached via the outer loop
performing poll(), so we can rewrite things to be easier to read by
just doing a larger read(), while at the same time ensuring that we
only report errors when errno is actually reliable.

Fixes: e0d32468
Signed-off-by: Eric Blake <eblake@redhat.com>