curl: Implement TCP keepalives and disabling Nagle’s algorithm.
Keepalives are useful for TCP connections to HTTP servers that may be
open for a long time (and if they fail, bring down the NBD connection
unless nbdkit-retry-filter is used).  Disabling Nagle’s algorithm is
not too likely to be useful since NBD request packets are generally at
least 512 bytes, but implementing these two together was easy.