server: Sanitize stdin/out before running plugin code
As shown in the previous patch, plugins may choose to use stdin or
stdout during .config.  But from .get_ready onwards, well-written
plugins shouldn't be needing any further use of stdin/out.  We already
swapped stdin/out to /dev/null while daemonizing, but did not do do
during -f or --run, which leads to some surprising inconsistency when
trying to debug a plugin that works in the foreground but fails when
daemonized.  What's more, while the task executed by --run should use
the original stdin/out (and we even have tests that would fail without
this), we don't want the plugin to interfere with whatever --run is
doing.  So it's time to move the swap over to /dev/null into a central
location, right before .get_ready.

Note that if a plugin consumes some, but not all, input during
.config, then we want --run to pick up with reading the rest of the
input.  For that to happen, we need to flush any buffered read-ahead
data from stdin.  POSIX says fflush(NULL) is supposed to be sufficient
for the task, but in practice, glibc still has a bug[1] that requires
an additional fflush(stdin).

[1] https://sourceware.org/bugzilla/show_bug.cgi?id=12799

Signed-off-by: Eric Blake <eblake@redhat.com>