eval, sh: Set $tmpdir before running the command, instead of globally.
The $tmpdir environment variable is used by the eval and sh plugins to
communicate the path to the temporary directory created by the plugin
for shell scripts to use for temporary files.

Previously we set this environment variable globally (in .load())
using setenv, but this means the environment variable can be leaked
into undesirable places, eg. into the --run script:

  $ nbdkit sh - --run 'echo tmpdir: $tmpdir' </dev/null
  tmpdir: /tmp/nbdkitshco6MC9

Instead of this we set it in a private copy of environ, so it is not
visible to the rest of the process.

After this change:

  $ ./nbdkit sh - --run 'echo tmpdir: $tmpdir' </dev/null
  tmpdir: