Add insert function and use the new vector library in several places.
This extends the vector library with an insert function.  It is more
expensive than appending, but this does not affect existing code using
vector and can be used in new places without making those uses more
expensive.

We use this function in nbdkit-extentlist-filter, nbdkit-eval-plugin
and the sparse library.

This is refactoring, so should not affect functionality at all.