server/locks: Allow lock_request to be called when there is no current conn.
On Haiku tests/test-socket-activation failed with:

  nbdkit: locks.c:96:lock_request: conn != NULL

called from server/sockets.c: accept_connection() in the fallback path
which does:

  lock_request ();
  thread_data->sock = set_cloexec (accept (listen_sock, NULL, NULL));
  unlock_request ()

lock_request assert fails because there is no current connection at
this point (as we are just at the point of accepting a new connection).

It should be possible to call lock_request without a connection,
provided that thread_model <= NBDKIT_THREAD_MODEL_SERIALIZE_REQUESTS
(which it is when the fallback path is engaged on Haiku).  This was a
regression when I modified the code to use an implicit thread-local
connection.

Fixes commit 91023f269d4cea56f573a1aa0d880b12052f6e1e.