tests/test-cache-max-size.sh: Fix use of ‘cut’ when date contains spaces.
This test fails only on the 1st-9th of each month.  As it’s the 1st May
today, the test is failing.

When searching for the file descriptor in /proc/<pid>/fd we grep for a
line which looks like this example:

  lrwx------. 1 rjones rjones 64 May  1 10:58 5 -> [...]

We used to use ‘awk’ to extract the file descriptor from the 9th field
(expecting 5 in this case).  However commit b927b47c61 switched to
using ‘cut’.  ‘cut’ does not squeeze multiple whitespace between
fields.  Since the date field today is “May<space><space>1” we now
extract the time (10:58) instead of the file descriptor, and the rest
of the test fails.

The solution is to use ‘tr’ to squeeze whitespace before cutting
(https://stackoverflow.com/a/4483833).

Fixes: commit b927b47c619a16f0d77a7d6461c66be2e11d7a4a.