vddk: Suppress errors in can_extents test (RHBZ#1709211).
In can_extents we test whether the VDDK function VixDiskLib_-
QueryAllocatedBlocks works before declaring that extents are
supported.  This VDDK function can fail in several ways but most
notably if the server doesn't support querying extents.

In this case VDDK raises an error message through the error_function
callback, which we turn into a call to nbdkit_error.  In most cases
this will cause something to be printed to stderr and/or logged
noisily to syslog.  This is undesirable because these are not errors.

This commit adds a per-thread error suppression mechanism so that we
can temporarily hide errors, especially while doing this test, but I
guess it might be useful in future in other cases too.

Note this only affects a narrow range of versions of VMware ESXi and
VDDK.  In particular to see this problem you would need to be using
VMware ESXi (server) <= 5.5 and VDDK (client) >= 6.7 at the same time.