eval: Fix free method_scripts vector in unload callback.
When freeing the method scripts this attempted to free the whole
struct (which is part of an array of structs) not the .script member
of the struct.  This was a regression introduced when rewriting this
plugin to use the vector type.

Fixes commit b05ad58fcdb58bd1a541284d1e76a3b9a0a91e0b.

(cherry picked from commit 5d834c900bf0c415bc628cf997ffe6567de330aa)