tests/test-truncate4.sh: Rewrite to use nbdsh.
The old test relied on issuing an asynchronous command to a qemu-io
subprocess, requiring a sleep to ensure that the command had been
carried out.  ‘sleep 1’ was not long enough on s390x as you can see
from this partial trace:

    + exec
    + qemu-io -f raw nbd:unix:/tmp/tmp.jqOZGj6dnX
    + exec
    + echo 'Reading from connection A, try 1'
    Reading from connection A, try 1
    + echo 'r -P 1 0 1024'
    + sleep 1

Above we issue the command and sleep while it executes.  However:

    nbdkit: debug: accepted connection
    nbdkit: file[1]: debug: truncate: preconnect
    nbdkit: file[1]: debug: file: preconnect
    nbdkit: file[1]: debug: newstyle negotiation: flags: global 0x3
    + echo 'Resizing down'
    Resizing down
    + truncate -s 512 truncate4.data
    nbdkit: file[1]: debug: newstyle negotiation: client flags: 0x3
    nbdkit: file[1]: debug: newstyle negotiation: NBD_OPT_STRUCTURED_REPLY: client requested structured replies

Here are still connecting to nbdkit from "connection A", long after
the sleep has finished and the truncate has been done.

Instead of increasing the sleep and hoping for the best, I chose to
rewrite the test to use nbdsh.  This allows us to control both
connections from a single process, making the test predicatable and
easier to understand.