vddk: Add undocumented noreexec=true option.
This undocumented option disables reexec.  The caller must set
LD_LIBRARY_PATH correctly as for older versions of the plugin.

This is intended for use only when debugging problems with reexec.  In
fact I added it because I suspected that reexec was causing a problem
with a particular version of VDDK.  In fact that did not turn out to
be the case.  But the option may still be useful in future for ruling
out reexec as a cause for issues.