vddk: Set connect params specType field correctly.
Our code was not setting the specType field properly, so initialize
it.  This probably won't have any effect because the field is
initialized to 0, but better safe than sorry.

Note this field was added in VDDK 6.7 along with the
AllocateConnectParams call.  Earlier versions of VDDK should simply
ignore the data past the end of the struct, which in those earlier
versions we are allocating.