common/allocators/sparse.c: Integrate back some changes from zstd.c
zstd.c was derived from sparse.c, and changed.  Integrate some of
those changes back into the original.

The main improvement is to define a struct (with a single entry) for
the L2 page entries, simplifying type declarations.

This is refactoring.  It does not change the implementation at all.

(cherry picked from commit 1ecfe860addfbb28d027c68807e1aeb0863c5254)

Note for stable-1.20 branch: “zstd” refers to the new memory
compression allocator added in the 1.21 development cycle.  This
commit backports some of the non-feature refactoring changes to the
stable branch.