tests/test-ssh.sh: Skip test if test disk is not created.
This test failed if mke2fs was too old to have the -d option (or
mke2fs was not installed at all), which means that the test disk was
not created.  Skip the test instead of failing in this case.

(cherry picked from commit 213ae82187e5b981b21019f00008d49a483d4a45)