Add Bittorrent plugin.
This contrived plugin lets you turn a Bittorrent download into an NBD
export.  One use case is to boot Linux installers without downloading
them (like nbdkit-curl-plugin, but with more fairness towards the file
hosting).  For example:

  $ wget https://cdimage.debian.org/debian-cd/current/amd64/bt-dvd/debian-10.4.0-amd64-DVD-1.iso.torrent
  $ nbdkit -U - torrent debian-10.4.0-amd64-DVD-1.iso.torrent \
           --run 'qemu-system-x86_64 -m 2048 -cdrom $nbd -boot d'

This should actually get faster and fairer the more people use it,
making it possibly useful as a way to distribute large disk images in
certain environments.

It requires libtorrent-rasterbar.  Note there are two different
"libtorrent" projects, the one we are using is located upstream at
https://www.libtorrent.org and packaged in Fedora as ‘rb_libtorrent’
and in Debian as ‘libtorrent-rasterbar’.