server: Add .after_fork callback, mainly for plugins to create threads.
If you have a plugin which either creates background threads itself or
uses a library that creates background threads, it turns out you
cannot create these in .get_ready (or earlier).  The reason is that
nbdkit forks when either daemonizing itself or using the --run option,
and fork invalidates all the background threads in the child process
(the daemonized or captive nbdkit).

The only good solution here is to add a new callback which I've called
.after_fork which runs after either daemonization or the --run
function.

It should be used sparingly because it's not possible to report errors
nicely from this callback.  In particular when daemonizing we have
lost the controlling terminal and errors go to syslog - it looks to
the user as if nbdkit “died”.

This adds the callback to the C and OCaml APIs, and also sh and eval
(to make the tests work), but I didn't update the other language APIs
yet.