nbd: Fix shared=true so it creates background thread after fork.
This option didn't work except when using --foreground (or using
another option that implies this).  This is because it creates a
background reader thread, but that thread is invalidated if nbdkit
forks, resulting in deadlocks because the worker threads are
fruitlessly sending notifications to a background thread that no
longer exists.

The fix is to move this work to .after_fork.