extents: Add nbdkit_extents_aligned()
We have several filters that would benefit from reporting extents to
the client that are always aligned to boundaries chosen by the filter,
regardless of whether the plugin reports extents at a finer
granularity.  Add a new helper function to make the work easier,
without having to duplicate code in each filter.  Any alignment block
that straddles more than one plugin extent is reported as a single
extent to the client, with a type determined by the intersection of
the underlying extents.

Signed-off-by: Eric Blake <eblake@redhat.com>