vector: Add VECT_remove
An upcoming patch wants to remove an arbitrary element from a vector.

Also, add testsuite coverage for other functions added since the
original unit test was written.  It's a bit awkward that the compare
for VECT_search and VECT_sort differ in type, but such is life (we
indeed have search code where typing the key differently is useful).

Signed-off-by: Eric Blake <eblake@redhat.com>