tests: Avoid odd test behavior under NDEBUG
While we support compilation with CFLAGS=-DNDEBUG for the brave user
desiring to avoid assertion overhead, this causes the tests to be less
powerful (the tests "fail open" for anything where we relied on an
assert to validate nbdkit was behaving correctly; and this would be
even worse if we had side effects in a test assert, such as we just
fixed in d5cedf20).  Fortunately, a quick test of './configure
CFLAGS=-DNDEBUG' didn't hit any test failures, but it seems better to
just ensure that assertions always work in our tests, even when they
are turned off for the main binary.

Signed-off-by: Eric Blake <eblake@redhat.com>