swab: Re-enable .extents
This reverts commit 2c5aec42cf04c567639bd885cf12a57192229215.

Now that we have an easy way to generate aligned extents from the
plugin, where the client will never see an unaligned extent
transition, it's time to put it to use.

The testsuite is an interesting mix of the eval plugin and --run
nbdsh, where I had to figure out how to pass various patterns in via a
shell script in extents= as well as parse them back out in a sane
manner in python.

Signed-off-by: Eric Blake <eblake@redhat.com>