New filter: gzip
Turn the existing nbdkit-gzip-plugin into a filter so it can be
applied on top of files or other sources:

  nbdkit file --filter=gzip file.gz
  nbdkit curl --filter=gzip https://example.com/disk.gz

Because of the nature of the gzip format which is not blocked based
and thus not seekable, this filter caches the whole uncompressed file
in a hidden temporary file.  This is required in order to implement
.get_size.  See this link for a more detailed explanation:
https://stackoverflow.com/a/9213826

This commit deprecates nbdkit-gzip-plugin and suggests removal in
nbdkit 1.26.