server: Deprecate the -e/--exportname parameter.
This parameter provided a name for the "default export" -- ie. the one
and only export returned to the client by NBD_OPT_LIST.  But nbdkit
traditionally didn't care what export name the client requested.
Since 1.16 plugins have been able to serve different content per
export name (and return errors for unknown exports), but the -e option
didn't reflect that and only created confusion.

What we actually have to do is ask the plugin what exports it provides
and return that list to the client.  This is for future work.

This commit deprecates and hides the parameter.  If used the option
now does nothing at all.

There are some visible but very minor changes resulting from this
commit:

(1) NBD_OPT_LIST advertises a single export called "" (instead of the
    -e parameter).  We intend to replace this implementation soon with a
    correct one.

(2) The --run option no longer sets $exportname to the -e parameter,
    but for backwards compat sets exportname=''.  It does not put
    export name into the $uri.  This was always the wrong thing to do
    since export names are per connection, not per server.  Existing
    --run scripts will see $exportname expand to '' which is most
    likely what they saw before and overwhelmingly more likely to be
    correct than if the -e option had been used.

(3) The -e option had the side-effect of forcing the newstyle
    protocol, so weirdly this worked and made the server use newstyle:

      nbdkit -o -e '' ...

    but this would fail with an error at start-up:

      nbdkit -e '' -o ...

    I thought it best to remove this odd behaviour completely.

(4) I have temporarily disabled tests/test-long-name.sh.  This is
    still a valid test, but we will have to rewrite it to use
    (probably) sh or eval plugins once we have an implementation of
    list_exports.