server: Pass the export name through filter .open calls.
To allow filters to modify the export name as it passes through the
layers this commit makes several changes:

The filter .open callback now takes an extra parameter, the export
name.  This is always non-NULL (for oldstyle it is "").  This string
has a short lifetime and filters that need to hang on to it must take
a copy.  The filter must pass the exportname parameter down to the
next layer (or potentially modify it).

The plugin .open callback is unchanged, but the binding to it in
server/plugins.c takes a copy of the exportname in the handle and this
is what nbdkit_export_name returns to plugins.  Eventually we will
deprecate that function and the V3 .open callback in plugins will have
the extra parameter, but this is not implemented yet.

nbdkit_export_name can only be called from plugins, not filters.
Filters should use the .open(...exportname) parameter if they need the
export name.

Filter .reopen also takes the exportname.  The only filter that uses
it (retry) must save the exportname from .open in order to pass it to
.reopen.  This filter already does the same for the readonly flag so
this seems reasonable.

The handling of NBD_OPT_SET_META_CONTEXT interacted with the export
name (see commit 20b8509a9ccdab118ce7b7043be63bbad74f1e79).  I have
attempted to keep previous behaviour in this change, but note that
there is no regression test for this.  I added a debug message so we
can tell when this unusual case actually happens which should help
with diagnosis of problems.

All of the above is internal refactoring and should have no visible
external effect on server behaviour or how plugins are written.
Filters need to be updated as discussed above.  All the in-tree
filters have been, and we assume there are no out of tree filters
because of the unstable filter API.

See also this thread:
https://www.redhat.com/archives/libguestfs/2020-July/thread.html#00087