streaming: Test that nbdkit + streaming + qemu-img really works.
The manual page for this plugin has long claimed that you can use it
to pipe data from qemu-img convert, if writing a raw file.  However
this was never routinely tested and is not self-evidently true, since
qemu-img might decide to write non-linearly even if writing raw.

This commit adds a test, proving that currently the claim _is_ true
(qemu-5.0.0-2.fc33.x86_64).  It will tell us if that assumption breaks
in future.