server: Add exports list functions
Add exports.c for manipulating an exports list, borrowing heavily from
similar code in managing an extents list.  The new functions are
exposed through nbdkit-filter.h, because filters will eventually need
a way to grab the export list from a plugin, but actually wiring that
up will be in a later patch.  For now, we enforce string length but
not strict UTF-8 content.

Signed-off-by: Eric Blake <eblake@redhat.com>