server: Prepare to use export list from plugin
Wire up everything in the server to query the export list from the
plugin.  Actual patches to let plugins and filters return a
non-default list will come later, so for now, the behavior is
unchanged: NBD_OPT_LIST still lists just "", and a client requesting
"" still lets the plugin see an export name of "".

If the client calls NBD_OPT_LIST, we cache the default export name for
later use in case the client then requests export "", to avoid having
to re-call .list_exports a second time.  But there is no caching if
the client calls NBD_OPT_LIST more than once, and we still have to
resolve the default name even if the client does not use NBD_OPT_LIST.

Signed-off-by: Eric Blake <eblake@redhat.com>