tests: Yet another test-nozero tweak
We discovered [1] that on xfs, even though the size of nozero2.img did
not change, the allocation increased when we wrote literal zeroes all
the way to end-of-file (that is, the filesystem presumed that we were
likely to write more data soon).  Try a couple more tricks to help
this test pass on koji: double the file size to 2M but keep the
write-zero operation at 1M (so that we are no longer writing all the
way to EOF), and relax the comparison to use -lt rather than !=
(although a growing file is odd, it is not fatal becuase we really
only care that the write zero did not become sparse).

[1] https://www.redhat.com/archives/libguestfs/2020-August/msg00017.html

Signed-off-by: Eric Blake <eblake@redhat.com>