tests: Another robustness tweak to test-nozero
Instead of looping to create the file, we can change our printf
argument to do it in one shot.  Since we can't control when the
filesystem might overallocate, it's easier to just compare a file to
its former size, rather than to try to compare between two files that
might have started with different allocated sizes even though they
have the same content; this in turn is made easy with a bash array.
In a few places, try to run more of the test rather than giving up at
the first error.

Signed-off-by: Eric Blake <eblake@redhat.com>