server: Call .get_ready before redirecting stdin/stdout to /dev/null.
VDDK plugin + --run was broken because of the following sequence of
events:

(1) After .config_complete, server redirects stdin/stdout to /dev/null.

(2) Server then calls vddk_get_ready which reexecs.

(3) We restart the server from the top, but with stdin/stdout
    redirected to /dev/null.  So saved_stdin/saved_stdout save
    /dev/null.

(4) run_command is called which "restores" /dev/null as stdin/stdout.

(5) The output of the --run option is sent to /dev/null.

In addition to fixing this problem with VDDK, it also makes general
sense not to redirect stdin/stdout before calling .get_ready since
this callback is supposed to be the last chance before the server
daemonizes, and redirecting stdin/stdout is part of daemonization.

Before this commit:

  $ ./nbdkit vddk libdir=tests/.libs /dev/null --run 'qemu-img info $nbd'
  [no output]

After this commit:

  $ ./nbdkit vddk libdir=tests/.libs /dev/null --run 'qemu-img info $nbd'
  image: nbd://localhost:10809
  file format: raw
  virtual size: 512 KiB (524288 bytes)
  disk size: unavailable

One test (test-stdio.sh) made the assumption that stdin/stdout had
already been redirected to /dev/null in .get_ready.  I adjusted this
test so it checks for redirection in .after_fork instead.