python: Acquire and release the Python GIL.
With the current thread model this doesn't matter since nbdkit itself
essentially enforces something like the GIL.  But if we want to
parallelize the Python plugin we must actually start using the GIL
right.

This is both complex and very poorly documented (including a lot of
plainly incorrect information online).  However I believe what I have
written here is right, and it appears to work.