server: Implement nbdkit_is_tls for use during .open
Now that we can differentiate content based on export name, we also
need to be able to differentiate content for a --tls=on server, since
TLS is optional according to whether the client has authenticated.

For internal code and filters, this means adding a new parameter; the
sh plugin can do likewise.  For plugins, we can't add a parameter
until the V3 protocol, so in the meantime, we add nbdkit_is_tls(),
even though it will be deprecated alongside nbdkit_export_name() when
we do get to V3.

Signed-off-by: Eric Blake <eblake@redhat.com>