tls-fallback: New filter
Take advantage of the fact that we can now detect the type of client
during --tls=on in order to provide safe dummy content for plaintext
clients without having to rewrite plugins to do so.

The filter isn't quite perfect yet: we need to fix nbdkit to allow
.list_exports to know when tls is in use, as well as to allow us to
completely skip the plugin's next_open().  But those can be followup
patches, hopefully before we stabilize the 1.22 release.

Signed-off-by: Eric Blake <eblake@redhat.com>