server: Expose final thread_model to filter's .get_ready
The next patch wants to add a filter that will prevent DoS attacks
from a plaintext client; to be successful, the filter must guarantee
that nbdkit did not settle on SERIALIZE_CONNECTIONS.  The easiest way
to solve this is to expose the final thread model to .get_ready, which
is after the point where .config_complete may have altered it, and
before any connections are permitted.

For now, we are not exposing this information to plugins.  In fact, if
we ever add new thread models, we would have to document that a plugin
must be prepared to end up in a thread model that was not known at
compile time (necessarily a mode tighter than the maximum the plugin
supports).  But if it turns out a v2 plugin needs it, we could add
nbdkit_get_thread_model, similar to nbdkit_is_tls or
nbdkit_export_name.

Signed-off-by: Eric Blake <eblake@redhat.com>